+++ lib/fastlane/plugin/wpmreleasetoolkit/helper/android/android_localize_helper.rb
@@ -254,7 +254,7 @@ module Fastlane
             all_xml_documents = glotpress_filters.map do |filters|
               UI.message "Downloading translations for '#{lang_codes[:android]}' from GlotPress (#{lang_codes[:glotpress]}) [#{filters}]..."
               download_glotpress_export_file(project_url: glotpress_project_url, locale: lang_codes[:glotpress], filters: filters)
-            end.compact
+            end
             next if all_xml_documents.empty?
 
             # Merge all XMLs together